+++ {{cookiecutter.project_slug}}/{{cookiecutter.project_slug}}/users/tests/test_models.py
@@ -1,4 +1,5 @@
 import pytest
+from hamcrest import assert_that, equal_to, is_
 
 from django.conf import settings
 
@@ -6,4 +7,5 @@ pytestmark = pytest.mark.django_db
 
 
 def test_user_get_absolute_url(user: settings.AUTH_USER_MODEL):
-    assert user.get_absolute_url() == '/users/{0}/'.format(user.username)
+    expected_url = '/users/{0}/'.format(user.username)
+    assert_that(user.get_absolute_url(), is_(equal_to(expected_url)))